+++ src/main/java/icu/samnyan/aqua/sega/chunithm/service/UserActivityService.java
@@ -35,10 +35,10 @@ public class UserActivityService {
     }
 
     public List<UserActivity> getAllByUserIdAndKind(String userId, String kind) {
-        return userActivityRepository.findAllByUser_Card_ExtIdAndKindOrderBySortNumberDesc(Integer.parseInt(userId), Integer.parseInt(kind));
+        return userActivityRepository.findAllByUser_Card_ExtIdAndKindOrderBySortNumberDesc(Long.parseLong(userId), Integer.parseInt(kind));
     }
 
     public List<UserActivity> getByUserId(String userId) {
-        return userActivityRepository.findAllByUser_Card_ExtId(Integer.parseInt(userId));
+        return userActivityRepository.findAllByUser_Card_ExtId(Long.parseLong(userId));
     }
 }
