+++ src/main/java/icu/samnyan/aqua/sega/chunithm/service/UserGameOptionService.java
@@ -30,6 +30,6 @@ public class UserGameOptionService {
     }
 
     public Optional<UserGameOption> getByUserId(String userId) {
-        return userGameOptionRepository.findByUser_Card_ExtId(Integer.parseInt(userId));
+        return userGameOptionRepository.findByUser_Card_ExtId(Long.parseLong(userId));
     }
 }
