+++ src/main/java/icu/samnyan/aqua/sega/ongeki/dao/userdata/UserActivityRepository.java
@@ -15,11 +15,11 @@ import java.util.Optional;
 @Repository("OngekiUserActivityRepository")
 public interface UserActivityRepository extends JpaRepository<UserActivity, Long> {
 
-    List<UserActivity> findByUser_Card_ExtId(int userId);
+    List<UserActivity> findByUser_Card_ExtId(long userId);
 
     Optional<UserActivity> findByUserAndKindAndActivityId(UserData userData, int kind, int activityId);
 
-    List<UserActivity> findByUser_Card_ExtIdAndKindOrderBySortNumberDesc(int userId, int kind);
+    List<UserActivity> findByUser_Card_ExtIdAndKindOrderBySortNumberDesc(long userId, int kind);
 
     @Transactional
     void deleteByUser(UserData user);
